requirements.txt
42cc-pystyle==0.0.15
Django==1.6.7
flake8==2.4.1
mccabe==0.3.1
pep8==1.5.7
Pillow==2.9.0
pyflakes==0.8.1
South==0.8.4
uWSGI==2.0.11.1
wheel==0.24.0
mysql-python

